Move routes to index files and update navbar
