+++ src/routes/_layout.svelte
@@ -2,11 +2,11 @@
   import Navbar from "../components/Navbar.svelte";
   // bunch of css import for various modules
   import "tabulator-tables/dist/css/tabulator.min.css";
-  //   import "bootstrap";
   import "bootstrap/dist/css/bootstrap.min.css";
+  export let segment;
 </script>
 
-<Navbar />
+<Navbar {segment} />
 <main>
   <slot />
 </main>
